# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.6.2 🍺
 * Do not edit manually.
 * Api
 * Elemoria Rise of the Pokemon API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Player {
  id: number;
  name: string;
  avatarColor: string;
  level: number;
  experience: number;
  currency: number;
  currentRegionId: number;
  totalBattles: number;
  battlesWon: number;
  totalPokemonCaptured: number;
  badges: string[];
  createdAt: string;
}

export interface CreatePlayerBody {
  name: string;
  avatarColor: string;
  starterPokemonpeciesId: number;
}

export interface UpdatePlayerProgressBody {
  experience?: number;
  currency?: number;
  currentRegionId?: number;
  totalBattles?: number;
  battlesWon?: number;
  totalPokemonCaptured?: number;
  badges?: string[];
}

export interface GameSummary {
  totalPlayers: number;
  totalPokemonCaptured: number;
  totalBattlesFought: number;
  totalSpeciesDiscovered: number;
  topPlayerName: string;
  topPlayerLevel: number;
}

export interface Move {
  id: number;
  name: string;
  type: string;
  power: number;
  accuracy: number;
  pp: number;
  category: string;
  description: string;
  effect?: string | null;
}

export interface Pokemonpecies {
  id: number;
  name: string;
  description: string;
  type1: string;
  type2?: string | null;
  baseHp: number;
  baseAttack: number;
  baseDefense: number;
  baseSpeed: number;
  baseSpecialAttack: number;
  baseSpecialDefense: number;
  evolutionLevel?: number | null;
  evolvesIntoId?: number | null;
  catchRate: number;
  spriteUrl: string;
  backSpriteUrl: string;
  moves: Move[];
  rarity: string;
}

export interface PlayerPokemon {
  id: number;
  playerId: number;
  speciesId: number;
  nickname?: string | null;
  level: number;
  experience: number;
  currentHp: number;
  maxHp: number;
  attack: number;
  defense: number;
  speed: number;
  specialAttack: number;
  specialDefense: number;
  moves: Move[];
  isInParty: boolean;
  partySlot?: number | null;
  statusCondition?: string | null;
  species: Pokemonpecies;
  capturedAt: string;
}

export interface CapturePokemonBody {
  speciesId: number;
  level: number;
  nickname?: string;
}

export interface UpdatePlayerPokemonBody {
  nickname?: string;
  currentHp?: number;
  experience?: number;
  level?: number;
  statusCondition?: string | null;
  evolveToSpeciesId?: number;
}

export interface SetPartyBody {
  /** @maxItems 6 */
  PokemonIds: number[];
}

export interface BattlePokemon {
  playerPokemonId?: number | null;
  speciesId: number;
  name: string;
  level: number;
  currentHp: number;
  maxHp: number;
  moves: Move[];
  statusCondition?: string | null;
  spriteUrl: string;
  backSpriteUrl: string;
  type1: string;
  type2?: string | null;
}

export interface BattleState {
  id: number;
  type: string;
  status: string;
  turn: number;
  playerPokemon: BattlePokemon;
  opponentPokemon: BattlePokemon;
  playerParty: BattlePokemon[];
  requiresSwitch: boolean;
  log: string[];
  canCapture: boolean;
  result?: string | null;
  expGained?: number | null;
  currencyGained?: number | null;
}

export type StartBattleBodyBattleType =
  (typeof StartBattleBodyBattleType)[keyof typeof StartBattleBodyBattleType];

export const StartBattleBodyBattleType = {
  wild: "wild",
  trainer: "trainer",
} as const;

export interface StartBattleBody {
  battleType: StartBattleBodyBattleType;
  regionId: number;
  trainerLevel?: number;
}

export type BattleActionBodyAction =
  (typeof BattleActionBodyAction)[keyof typeof BattleActionBodyAction];

export const BattleActionBodyAction = {
  attack: "attack",
  item: "item",
  flee: "flee",
  capture: "capture",
  switch: "switch",
} as const;

export interface BattleActionBody {
  action: BattleActionBodyAction;
  moveId?: number;
  itemId?: number;
  switchToPokemonId?: number;
}

export interface BattleRecord {
  id: number;
  type: string;
  result: string;
  opponentName: string;
  expGained: number;
  currencyGained: number;
  createdAt: string;
}

export interface Region {
  id: number;
  name: string;
  description: string;
  minLevel: number;
  maxLevel: number;
  biome: string;
  isUnlocked: boolean;
  unlockRequirement?: string | null;
  imageUrl: string;
}

export interface WildPokemonEncounter {
  speciesId: number;
  species: Pokemonpecies;
  minLevel: number;
  maxLevel: number;
  encounterRate: number;
}

export interface InventoryItem {
  id: number;
  itemType: string;
  name: string;
  description: string;
  quantity: number;
  effect: string;
  imageUrl: string;
}

export interface Inventory {
  items: InventoryItem[];
  totalSlots: number;
  usedSlots: number;
}

export interface AddInventoryItemBody {
  itemType: string;
  name: string;
  quantity: number;
}

export interface LeaderboardEntry {
  rank: number;
  playerName: string;
  level: number;
  battlesWon: number;
  totalPokemonCaptured: number;
  badges: string[];
}

export type HealParty200 = {
  success: boolean;
};