# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Elemoria Rise of the Pokemon API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /battle/start:
    post:
      operationId: startBattle
      tags: [battle]
      summary: Start a new battle (wild or trainer)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StartBattleBody"
      responses:
        "201":
          description: Battle started
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BattleState"
  /battle/{battleId}/action:
    post:
      operationId: performBattleAction
      tags: [battle]
      summary: Perform an action in battle (attack, use item, flee, switch)
      parameters:
        - name: battleId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BattleActionBody"
      responses:
        "200":
          description: Battle state after action
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BattleState"
  /battle/{battleId}:
    get:
      operationId: getBattle
      tags: [battle]
      summary: Get current battle state
      parameters:
        - name: battleId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Battle state
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BattleState"
...
components:
  schemas:
...
    Move:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        type:
          type: string
        power:
          type: integer
        accuracy:
          type: integer
        pp:
          type: integer
        category:
          type: string
        description:
          type: string
        effect:
          type: string
          nullable: true
      required:
        - id
        - name
        - type
        - power
        - accuracy
        - pp
        - category
        - description
...
    BattleState:
      type: object
      properties:
        id:
          type: integer
        type:
          type: string
        status:
          type: string
        turn:
          type: integer
        playerPokemon:
          $ref: "#/components/schemas/BattlePokemon"
        opponentPokemon:
          $ref: "#/components/schemas/BattlePokemon"
        playerParty:
          type: array
          items:
            $ref: "#/components/schemas/BattlePokemon"
        requiresSwitch:
          type: boolean
        log:
          type: array
          items:
            type: string
        canCapture:
          type: boolean
        result:
          type: string
          nullable: true
        expGained:
          type: integer
          nullable: true
        currencyGained:
          type: integer
          nullable: true
      required:
        - id
        - type
        - status
        - turn
        - playerPokemon
        - opponentPokemon
        - playerParty
        - requiresSwitch
        - log
        - canCapture
    BattlePokemon:
      type: object
      properties:
        playerPokemonId:
          type: integer
          nullable: true
        speciesId:
          type: integer
        name:
          type: string
        level:
          type: integer
        currentHp:
          type: integer
        maxHp:
          type: integer
        moves:
          type: array
          items:
            $ref: "#/components/schemas/Move"
        statusCondition:
          type: string
          nullable: true
        spriteUrl:
          type: string
        backSpriteUrl:
          type: string
        type1:
          type: string
        type2:
          type: string
          nullable: true
      required:
        - speciesId
        - name
        - level
        - currentHp
        - maxHp
        - moves
        - spriteUrl
        - backSpriteUrl
        - type1
    StartBattleBody:
      type: object
      properties:
        battleType:
          type: string
          enum: [wild, trainer]
        regionId:
          type: integer
        trainerLevel:
          type: integer
      required:
        - battleType
        - regionId
    BattleActionBody:
      type: object
      properties:
        action:
          type: string
          enum: [attack, item, flee, capture, switch]
        moveId:
          type: integer
        itemId:
          type: integer
        switchToPokemonId:
          type: integer
      required:
        - action
...